% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3.R
\name{dijkstra}
\alias{dijkstra}
\title{this is the solution for dijkstra Algorithm}
\usage{
dijkstra(wiki_graph, a)
}
\arguments{
\item{wiki_graph}{A DataFrame}

\item{a}{A Number}
}
\value{
A Vector
}
\description{
this is the solution for dijkstra Algorithm
}
