% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{euclidean algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric number.}

\item{b}{A numeric number.}
}
\value{
A numeric number.
}
\description{
euclidean algorithm is to caculate the greatest common divisor of two integers
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
wiki page link <https://en.wikipedia.org/wiki/Euclidean_algorithm>
}
