% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Function}
\usage{
euclidean(x, y)
}
\arguments{
\item{x, y}{respectively numerical values}
}
\value{
the greatest common divisor of two numbers
}
\description{
Euclidean Function
}
\section{Functions}{
\itemize{
\item \code{euclidean}: algoritm to implement is the Euclidian algorithm to nd the greatest common divisor of two numbers. The description of the algorithm with pseudocode can be found here https://en.wikipedia.org/wiki/Euclidean algorithm. Assert that the arguments are numeric scalars or integers.
}}

