% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_test.R
\name{gp_test}
\alias{gp_test}
\title{Gender against physical activity chi squared test}
\usage{
gp_test(data)
}
\arguments{
\item{data}{This is the input data to perform the test on}
}
\value{
List with a summary of the chi squared test and a table
}
\description{
This function performs a chi squared test of gender against physical activity on a dataset
}
\examples{
gp_test(project)
}
