% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg_test.R
\name{hg_test}
\alias{hg_test}
\title{Height against gender t test}
\usage{
hg_test(data)
}
\arguments{
\item{data}{This is the input data to perform the test on}
}
\value{
List with a summary of the t test and a histogram
}
\description{
This function performs a t test of height against gender on a dataset
}
\examples{
hg_test(project)
}
