% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw_test.R
\name{hw_test}
\alias{hw_test}
\title{Height to weight linear regression test}
\usage{
hw_test(data)
}
\arguments{
\item{data}{This is the input data to perform the test on}
}
\value{
List with a summary of the linear regression test and a graph
}
\description{
This function performs a linear regression test of height against weight on a dataset
}
\examples{
hw_test(project)
}
