% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\docType{data}
\name{project}
\alias{project}
\title{1000 observations of men and women aged 26-45}
\format{
A data frame with 1000 rows and 5 variables:
\describe{
\item{ID}{Identification number}
\item{gender}{Male or Female}
\item{height}{height in cm}
\item{weight}{weight in kg}
\item{phys}{amount of phyical acticty ("None", "Moderate", "Intense")}
...
}
}
\usage{
project
}
\description{
A dataset containing 1000 observations of men and women aged 26-45 with information on the following variables:
ID, Gender, Height, Weight, Physical activity
}
\keyword{datasets}
