% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapeoid.R
\name{trapezoid}
\alias{trapezoid}
\title{trapezoid rule}
\usage{
trapezoid(f, range, n)
}
\arguments{
\item{f}{a function with a single argument x}

\item{range}{a vector specifying the lower and upper end points of the range of integration (a and b)}

\item{n}{the number of intervals for the trapezoid approximation}
}
\value{
a single value which is the trapezoid approximation to the integral
}
\description{
trapezoid rule
}
\examples{
> trapezoid(function(x){x^2}, c(1,2), 100)
2.33335
}
