% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Title}
\format{
A data frame with nodes and weight:
\describe{
\item{v1}{this is the First Nodes v1}
\item{v2}{this is the 2nd Nodes v2}
\item{w}{this is the weight between Nodes}
}
}
\usage{
wiki_graph
}
\description{
Description
}
\keyword{datasets}
