\name{assignment3-package}
\alias{assignment3-package}
\alias{assignment3}
\docType{package}
\title{
\packageTitle{assignment3}
}
\description{
\packageDescription{assignment3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{assignment3}
\packageIndices{assignment3}
~~ An overview of how to use the package, including the most important ~~
~~ dijkstra is to find the shortest distince in a graph with a given start point ~~
~~ euclidean is to caculate the greatest common divisor of two integers ~~
}
\author{
\packageAuthor{assignment3}

Maintainer: \packageMaintainer{assignment3}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages ~~
}
