% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3.R
\name{euclidean}
\alias{euclidean}
\title{euclidean of two Numbers}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
A number.
}
\description{
euclidean of two Numbers
}
\examples{
euclidean(100,1000)
euclidean(123612, 13892347912)
}
